Move FolderPickerWithAdd implementation to FolderPicker
